requirements.txt
# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Background Tasks
celery==5.3.4
redis==5.0.1

# Security
bcrypt==4.1.2
python-dotenv==1.0.0

# Validation
marshmallow==3.20.1
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Development
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Production Server
gunicorn==21.2.0
gevent==24.2.1
